refactor: Create component for step visualization
