+++ src/algorithms/utils.py
@@ -6,6 +6,10 @@ import pandas as pd
 
 
 def get_samples(data, num_samples) -> List:
+    if isinstance(data, int):
+        array = np.arange(data)
+        np.random.shuffle(array)
+        return list(array[:num_samples])
     array = np.arange(data.shape[0])
     np.random.shuffle(array)
     return list(array[:num_samples])
