+++ src/widgets/components/__init__.py
@@ -1,0 +1,2 @@
+from .clustering_steps import ClusteringStepsTemplate
+from .parameters import ParametersGroupBox
